feat(functions/v2): use FF testing layer for GCS sample (#2646)
* updated GCS v2 sample to new FF testing layer

* fixed lint

* fixed lint

* updated copyright year

* Revert copyright date change

* Address comments
