+++ functions/v2/helloGCS/test/unit.test.js
@@ -1,4 +1,4 @@
-// Copyright 2021 Google LLC
+// Copyright 2022 Google LLC
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,29 +12,27 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-const assert = require('assert');
-const sinon = require('sinon');
-const supertest = require('supertest');
-
-const functionsFramework = require('@google-cloud/functions-framework/testing');
-
-beforeEach(() => {
-  // require the module that includes the functions we are testing
-  require('../index');
-
-  // stub the console so we can use it for side effect assertions
-  sinon.stub(console, 'log');
-  sinon.stub(console, 'error');
-});
-
-afterEach(() => {
-  // restore the console stub
-  console.log.restore();
-  console.error.restore();
-});
+const {getFunction} = require('@google-cloud/functions-framework/testing');
 
 describe('functions_cloudevent_storage', () => {
-  it('should process a CloudEvent', async () => {
+  const assert = require('assert');
+  const sinon = require('sinon');
+  require('..');
+
+  const stubConsole = function () {
+    sinon.stub(console, 'error');
+    sinon.stub(console, 'log');
+  };
+
+  const restoreConsole = function () {
+    console.log.restore();
+    console.error.restore();
+  };
+
+  beforeEach(stubConsole);
+  afterEach(restoreConsole);
+
+  it('helloGCS: should print out event', () => {
     const event = {
       id: '1234',
       type: 'mock-gcs-event',
@@ -43,12 +41,10 @@ describe('functions_cloudevent_storage', () => {
         name: 'my-file.txt',
       },
     };
-    const server = functionsFramework.getTestServer('helloGCS');
-    await supertest(server)
-      .post('/')
-      .send(event)
-      .set('Content-Type', 'application/json')
-      .expect(204);
+
+    // Call tested function and verify its behavior
+    const helloGCS = getFunction('helloGCS');
+    helloGCS(event);
 
     assert(console.log.calledWith('Event ID: 1234'));
     assert(console.log.calledWith('Event Type: mock-gcs-event'));
